fixed nodes distance indexing logic

+++ include/refinement.h
@@ -4,6 +4,6 @@
 #include <stdlib.h>
 #include "logger.h"
 
-void two_opt(double *matrix, int size_node, double *node);
+void two_opt(double *matrix, int size_node, int *node);
 
 #endif
